+++ sdk/highlight-next/CHANGELOG.md
@@ -75,3 +75,11 @@
 
 - Tune settings of opentelemetry SDK to reduce memory usage.
 - Enable GZIP compression of exported data.
+
+### 4.4.0
+
+### Minor changes
+
+- Repaired `Highlight` export from `@highlight-run/next/server` to keep serverless functions alive while flushing OTEL errors
+- Added `H` and `Highlight` exports from `@highlight-run/next/edge` to wrap Vercel's Edge Runtime API endpoints
+- Added `H` and `Highlight` exports from `@highlight-run/next/app-router` to wrap App Router API endpoints